requirements.txt
bcrypt==4.1.3
blinker==1.8.2
cachelib==0.13.0
certifi==2024.7.4
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-WTF==1.2.1
idna==3.7
is-disposable-email==1.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgspec==0.18.6
pymongo==3.12.0
python-dotenv==1.0.1
Werkzeug==3.0.3
WTForms==3.1.2
wtforms-validators==1.0.0
